requirements.txt
livekit-agents
livekit-server-sdk
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
